refactor(patina): split output into cohesive modules
